+++ R/spbabel-package.r
@@ -47,3 +47,10 @@ NULL
 #' @docType data
 #' @rdname holey
 NULL
+
+#' Multi-object track with x, y, z, and time. 
+#' Created in /data-raw/track.r
+#' @name track
+#' @docType data
+#' @rdname track
+NULL