+++ NEWS.md
@@ -1,5 +1,7 @@
 # spbabel dev
 
+* added track data set
+
 * added holey data set
 
 * update to use tibble rather than dplyr data_frame